tests/wpt/tests/html/semantics/scripting-1/the-script-element/execution-timing/028.html

<!DOCTYPE html>
<html><head>
        <title> scheduler: javascript: URL</title>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="testlib/testlib.js"></script>
</head>
<body>

        <div id="log">FAILED (This TC requires JavaScript enabled)</div>
        <script>log('inline script #1');
                window.location.replace('javascript:log(\'JS URL\')');
                log('end script #1');
        </script>

        <script type="text/javascript">
        log( 'inline script #2' );
        var t = async_test()

        function test() {
                assert_any(assert_array_equals, eventOrder, [
                  ['inline script #1', 'end script #1', 'JS URL', 'inline script #2'],
                  ['inline script #1', 'end script #1', 'inline script #2', 'JS URL']]);
                t.done();
        }
        onload = t.step_func(test)
        </script>

</body></html>
